backend/requirements.txt
# FastAPI Inventory Management - Backend Requirements
# =====================================================
# Python 3.10+ required

# Core Framework
fastapi==0.109.0              # Modern async web framework
uvicorn[standard]==0.27.0     # ASGI server for FastAPI
pydantic==2.5.3               # Data validation with type hints
pydantic-settings==2.1.0      # Settings management from env vars

# Database
aiomysql==0.2.0               # Async MySQL driver
PyMySQL==1.1.0                # MySQL driver (dependency of aiomysql)

# Authentication
python-jose[cryptography]==3.3.0  # JWT token creation and validation
passlib[bcrypt]==1.7.4        # Password hashing (for future use)
python-multipart==0.0.6       # Form data parsing

# Environment & Configuration
python-dotenv==1.0.0          # Load .env files

# Production Server (Optional)
gunicorn==21.2.0              # Production WSGI server

# Development Tools (Optional)
# pytest==7.4.4               # Testing framework
# httpx==0.26.0               # Async HTTP client for testing
# black==24.1.1               # Code formatter
# ruff==0.1.14                # Fast linter

